app/requirements.txt
libsass==0.22.0
mysqlclient==2.2.0
pytz==2023.3
PyYAML==6.0
uWSGI==2.0.21
